Avoid possible error where name does not include /

+++ library/KT/Person.php
@@ -1778,7 +1778,7 @@ class KT_Person extends KT_GedcomRecord {
 		// Add placeholder for unknown given name
 		if (!$GIVN) {
 			$GIVN	= '@P.N.';
-			$pos	= strpos($full, '/');
+			$pos	= (int) strpos($full, '/');
 			$full	= substr($full, 0, $pos) . '@P.N. ' . substr($full, $pos);
 		}
